pyproject.toml
[project]
name = "migracoes-venezuelanas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.4",
    "pydantic-ai-slim[google,mistral]>=1.27.0",
    "python-dotenv>=1.2.1",
    "streamlit-nightly>=1.51.1.dev20251124",
    "thefuzz>=0.22.1",
]
